fix: 数据库表破坏性错误导致同步异常
