+++ FlowDown.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -293,8 +293,8 @@
       "kind" : "remoteSourceControl",
       "location" : "https://github.com/huggingface/swift-transformers",
       "state" : {
-        "revision" : "18ccbd247eb4bf35a0ea98aab3b95ca255f09376",
-        "version" : "1.1.3"
+        "revision" : "296c0e901c6b37164af82c57bbdcc0d69190f5ce",
+        "version" : "1.1.4"
       }
     },
     {
