+++ Frameworks/Storage/Sources/Storage/Storage/DBVersion.swift
@@ -12,6 +12,4 @@ enum DBVersion: Int, CaseIterable {
     case Version3 = 3
     case Version4 = 4
     case Version5 = 5
-    case Version6 = 6
-    case Version7 = 7
 }
