+++ Frameworks/Storage/Sources/Storage/Storage/Storage.swift
@@ -83,8 +83,6 @@ public class Storage {
                 MigrationV2ToV3(),
                 MigrationV3ToV4(),
                 MigrationV4ToV5(),
-                MigrationV5ToV6(),
-                MigrationV6ToV7(),
             ]
         } else {
             initVersion = .Version1
@@ -93,8 +91,6 @@ public class Storage {
                 MigrationV2ToV3(),
                 MigrationV3ToV4(),
                 MigrationV4ToV5(),
-                MigrationV5ToV6(),
-                MigrationV6ToV7(),
             ]
         }
