+++ Frameworks/Storage/Sources/Storage/Tables/V1/CloudModelV1.swift
@@ -20,6 +20,7 @@ package final class CloudModelV1: Identifiable, Codable, Equatable, Hashable, Ta
     package var headers: [String: String] = [:] // additional headers
     package var capabilities: Set<ModelCapabilities> = []
     package var context: ModelContextLength = .short_8k
+    package var temperature_preference: ModelTemperaturePreference = .inherit
 
     // can be used when loading model from our server
     // present to user on the top of the editor page
@@ -38,6 +39,7 @@ package final class CloudModelV1: Identifiable, Codable, Equatable, Hashable, Ta
             BindColumnConstraint(capabilities, isNotNull: true, defaultTo: Set<ModelCapabilities>())
             BindColumnConstraint(context, isNotNull: true, defaultTo: ModelContextLength.short_8k)
             BindColumnConstraint(comment, isNotNull: true, defaultTo: "")
+            BindColumnConstraint(temperature_preference, isNotNull: true, defaultTo: ModelTemperaturePreference.inherit)
         }
 
         case id
@@ -50,6 +52,7 @@ package final class CloudModelV1: Identifiable, Codable, Equatable, Hashable, Ta
         case capabilities
         case context
         case comment
+        case temperature_preference
     }
 
     package init(
@@ -63,6 +66,7 @@ package final class CloudModelV1: Identifiable, Codable, Equatable, Hashable, Ta
         context _: ModelContextLength = .medium_64k,
         capabilities: Set<ModelCapabilities> = [],
         comment: String = "",
+        temperature_preference: ModelTemperaturePreference = .inherit,
     ) {
         self.id = id
         self.model_identifier = model_identifier
@@ -73,6 +77,7 @@ package final class CloudModelV1: Identifiable, Codable, Equatable, Hashable, Ta
         self.headers = headers
         self.capabilities = capabilities
         self.comment = comment
+        self.temperature_preference = temperature_preference
     }
 
     package required init(from decoder: Decoder) throws {
@@ -87,6 +92,7 @@ package final class CloudModelV1: Identifiable, Codable, Equatable, Hashable, Ta
         capabilities = try container.decodeIfPresent(Set<ModelCapabilities>.self, forKey: .capabilities) ?? []
         context = try container.decodeIfPresent(ModelContextLength.self, forKey: .context) ?? .short_8k
         comment = try container.decodeIfPresent(String.self, forKey: .comment) ?? ""
+        temperature_preference = try container.decodeIfPresent(ModelTemperaturePreference.self, forKey: .temperature_preference) ?? .inherit
     }
 
     package static func == (lhs: CloudModelV1, rhs: CloudModelV1) -> Bool {
@@ -104,5 +110,6 @@ package final class CloudModelV1: Identifiable, Codable, Equatable, Hashable, Ta
         hasher.combine(capabilities)
         hasher.combine(context)
         hasher.combine(comment)
+        hasher.combine(temperature_preference)
     }
 }
